requirements.txt
nltk
newspaper3k
python-telegram-bot
python-dotenv
